natssysclient: Extract request to a separate method, contextualize errors
