+++ natssysclient/statsz_server.go
@@ -16,11 +16,8 @@ package natssysclient
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"time"
-
-	"github.com/nats-io/nats.go"
 )
 
 type (
@@ -91,25 +88,12 @@ type (
 
 // ServerStatsz returns server statistics for the Statsz structs.
 func (s *System) ServerStatsz(ctx context.Context, id string, opts StatszEventOptions) (*ServerStatszResp, error) {
-	if id == "" {
-		return nil, fmt.Errorf("%w: server id cannot be empty", ErrValidation)
-	}
-	conn := s.nc
-	subj := fmt.Sprintf(srvStatszSubj, id)
 	payload, err := json.Marshal(opts)
 	if err != nil {
 		return nil, err
 	}
-	if _, ok := ctx.Deadline(); !ok {
-		var cancel context.CancelFunc
-		ctx, cancel = context.WithTimeout(ctx, DefaultRequestTimeout)
-		defer cancel()
-	}
-	resp, err := conn.RequestWithContext(ctx, subj, payload)
+	resp, err := s.requestByID(ctx, id, srvStatszSubj, payload)
 	if err != nil {
-		if errors.Is(err, nats.ErrNoResponders) {
-			return nil, fmt.Errorf("%w: %s", ErrInvalidServerID, id)
-		}
 		return nil, err
 	}
